Amend description of task creation; bullet list formatting
If I'm not mistaken, association of a task with a job is already done upon the task's creation.  This can be removed from the list of "more things that we need to add to the model".

Minor: the last entry doesn't require a semicolon; add "and" before the last item.