Add failing comparison test cases
The __eq__() comparison methods as added to the pyjobshop.ProblemData
module return True when comparing an object with its subclass.  This is
undesirable because the subclass may have extra attributes which are
then not checked for.

This (failing) test case adds a check for this case.  The test passes
from the next commit on.